backend/github_clone/requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
auto_dataclass==0.1.6
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==4.2.7
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.4
django-redis==5.0.0
djangorestframework==3.14.0
djangorestframework-dataclasses==1.3.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
Faker==20.1.0
gitdb==4.0.11
GitPython==3.1.42
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
kombu==5.3.4
packaging==23.2
pluggy==1.3.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-git==2018.2.1
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
Send2Trash==1.8.2
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
unidiff==0.7.5
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
zope.interface==6.3
